pyproject.toml
[tool.poetry]
name = "xdata-datascientist"
version = "0.1.0"
description = ""
authors = ["Joen Tan <joentze01@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
fastapi = "^0.135.3"
torch = "^2.11.0"
nemo-toolkit = {extras = ["asr"], version = "^2.7.2"}
librosa = "^0.10"
python-multipart = "^0.0.26"
uvicorn = "^0.44.0"
pydub = "^0.25.1"
jiwer = "3.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
